backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.95.0",
    "metatrader5>=5.0.46",
    "sqlmodel>=0.0.8",
    "uvicorn[standard]>=0.22.0",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
]
